% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_FXs.R
\name{AtlasRunSingleR}
\alias{AtlasRunSingleR}
\title{Modified Copy of CellMembrane's RunSingleR}
\usage{
AtlasRunSingleR(
  seuratObj = NULL,
  datasets = c("rhesusAtlas"),
  assay = NULL,
  resultTableFile = NULL,
  rawDataFile = NULL,
  minFraction = 0.01,
  showHeatmap = TRUE,
  maxCellsForHeatmap = 20000,
  nThreads = 8,
  sceAtlas = NULL
)
}
\arguments{
\item{seuratObj}{The Seurat Object to be labeled by SingleR}

\item{datasets}{One or more datasets to use as a reference. Allowable values are: hpca, blueprint, dice, monaco, rhesusAtlas, and immgen. See cellDex package for available datasets.}

\item{assay}{The assay in the seurat object to use}

\item{resultTableFile}{If provided, a table of results will be saved here}

\item{rawDataFile}{If provided, the complete SingleR results will be saved to this file}

\item{minFraction}{If provided, any labels present with fraction of this or fewer across cells will be converted to Unknown}

\item{showHeatmap}{If true, heatmaps will be generated showing the SingleR calls}

\item{maxCellsForHeatmap}{The heatmap will only be plotted if the total cells is below this number}

\item{nThreads}{If provided, this integer value is passed to SingleR's BPPARAM argument. On windows ths is passed to BiocParallel::SnowParam(). On other OS it is passed to BiocParallel::MulticoreParam()}
}
\value{
The modified seurat object
}
\description{
This copy allows for a local reference to be run, until we host the atlas remotely. Defaults to parallelization (8 cores) and the rhesusAtlas (to be provided from running ConvertAtlasAndApplyMetadata() ).
}
\examples{
seuratObj.test <- AtlasRunSingleR(seuratObj = seuratObj.test, datasets = c("rhesusAtlas"), sceAtlas = sceAtlas)
}
