% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModules.R
\name{CalculateUCellScores}
\alias{CalculateUCellScores}
\title{CalculateUCellScores}
\usage{
CalculateUCellScores(seuratObj, forceRecalculate = FALSE)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{forceRecalculate}{If true, UCell will always be re-run, even if the field is already present.}
}
\description{
This will calculate a handful of standardized UCell scores for a seurat object
}
