% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_FXs.R
\name{classifier_Seurat}
\alias{classifier_Seurat}
\title{classifier_Seurat}
\usage{
classifier_Seurat(
  Ref.integrated = NULL,
  Query.Ser = NULL,
  dims = 1:30,
  dims_umap = 1:20,
  col.name = NULL
)
}
\arguments{
\item{Ref.integrated}{The seurat object of the Ref.integrated}

\item{Query.Ser}{The seurat object of the Query}

\item{dims}{CCA/PCA dims to integrate default 1:30}

\item{dims_umap}{CCA/PCA dims to run UMAP default 1:20}

\item{col.name}{metadata col name pertaining to Ref to be mapped to Query}
}
\description{
calls seurat fx to classify a query dataset
}
