% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{Classify_ImmuneCells}
\alias{Classify_ImmuneCells}
\title{Classify Bulk Immune cells}
\usage{
Classify_ImmuneCells(
  seuratObj,
  assayName = Seurat::DefaultAssay(seuratObj),
  columnPrefix = "RIRA_Immune_v2.",
  maxAllowableClasses = 6,
  minFractionToInclude = 0.01,
  minCellsToRun = 200,
  maxBatchSize = 6e+05,
  retainProbabilityMatrix = FALSE,
  filterDisallowedClasses = TRUE,
  maxAllowedUnknown = 0.2
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{assayName}{The name of the assay to use. Others will be dropped}

\item{columnPrefix}{A prefix that will be added to the beginning of the resulting columns, added the seurat@meta.data}

\item{maxAllowableClasses}{Celltypist can assign a cell to many classes, creating extremely long labels. Any cell with more than this number of labels will be set to NA}

\item{minFractionToInclude}{If non-null, any labels with fewer than this fraction of cells will be set to NA.}

\item{minCellsToRun}{If the input seurat object has fewer than this many cells, NAs will be added for all expected columns and celltypist will not be run.}

\item{maxBatchSize}{If more than this many cells are in the object, it will be split into batches of this size and run in serial.}

\item{retainProbabilityMatrix}{If true, the celltypist probability_matrix with per-class probabilities will be stored in meta.data}

\item{filterDisallowedClasses}{If true, this will run FilterDisallowedClasses() on the output.}

\item{maxAllowedUnknown}{If provided, this method will throw an error if the fraction of cells assigned to Unknown is above this value}
}
\description{
Runs celltypist using the RIRA bulk immune model to score cells using CellTypist with optimized parameters.
}
