% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_FXs.R
\name{ConvertAtlasAndApplyMetadata}
\alias{ConvertAtlasAndApplyMetadata}
\title{Convert a Seurat object atlas into a Single Cell Experiment type Atlas and apply relevant metadata}
\usage{
ConvertAtlasAndApplyMetadata(
  seuratAtlas = NULL,
  label.main = "Pheno.scope1",
  label.fine = "Pheno.scope4"
)
}
\arguments{
\item{seuratAtlas}{A Seurat object type Rhesus Atlas}

\item{label.main}{The Pheno.scope that is used for the highest singleR scope. Default = Pheno.scope1}

\item{label.fine}{The Pheno.scope that is used for the lower singleR scope. Default = Pheno.scope4}
}
\value{
A SCE type atlas, ready to be used in SingleR functions.
}
\description{
Convert a Seurat object atlas into a Single Cell Experiment type Atlas and apply relevant metadata
}
\examples{
sceAtlas <- ConvertAtlasAndApplyMetadata(seuratAtlas)
sceAtlas <- ConvertAtlasAndApplyMetadata(seuratAtlas, label.main = "Pheno.scope2", label.fine = "Pheno.scope6")
}
