% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Wrappers.R
\name{DietSeurat_Cust}
\alias{DietSeurat_Cust}
\title{DietSeurat_Cust}
\usage{
DietSeurat_Cust(serobj = NULL, addCITE = F)
}
\arguments{
\item{serobj}{The seurat object where data will be added.}

\item{addCITE}{T/F}
}
\description{
a custom version of the DietSeurat() to really do the job
}
