% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{.FindGeneSetsContaining}
\alias{.FindGeneSetsContaining}
\title{.FindGeneSetsContaining}
\usage{
.FindGeneSetsContaining(genes)
}
\arguments{
\item{genes}{A vector of genes}
}
\description{
A helper function that Takes an input gene and identifies any entries matching registered gene sets. This function is the complement to GetGeneSet.
}
