% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.RemoveUCellDoublets}
\alias{.RemoveUCellDoublets}
\title{Remove UCell Doublets}
\usage{
.RemoveUCellDoublets(
  seuratObj,
  parentUCell = NULL,
  otherUCellScores = c("Platelet_MK_UCell", "Erythrocyte_UCell", "Neutrophil_UCell",
    "Epithelial_UCell", "Stromal_UCell"),
  clusterResolution = "ClusterNames_1.2",
  cosineSimilarityCutoff = 0.5
)
}
\arguments{
\item{parentUCell}{The subset's target cell type's UCell score. (Computed by scGate early in processing).}

\item{otherUCellScores}{A vector of cell type UCell scores to attempt to discard from downstream processing. By default, these are cell types that comprise the dump gate in RIRA's scGate model.}

\item{clusterResolution}{The metadata column name that defines cluster membership (e.g. ClusterNames_0.2)}

\item{cosineSimilarityCutoff}{A number between -1 and 1 that is used to threshold/gate the prospective doublets as defined by the computed score.}
}
\value{
Returns a seurat object with the UCell doublets removed.
}
\description{
Attempts to remove dump-gate "doublets" from the main three cell type subset objects (Tcell, Bcell, MoMacDC)
}
