% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentScores.R
\name{ExtractGeneWeights}
\alias{ExtractGeneWeights}
\title{ExtractGeneWeights}
\usage{
ExtractGeneWeights(
  seuratObj,
  componentNum,
  topNGenes = 100,
  bottomNGenes = 100,
  reduction = "pca"
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{componentNum}{The component number to use}

\item{topNGenes}{The number of top genes to use (ranked by loading)}

\item{bottomNGenes}{The number of bottom genes to use (ranked by loading)}

\item{reduction}{The reduction to use}
}
\description{
Extracts the loadings from a given component in a seurat object and returns a data.frame suitable for ScoreUsingSavedComponent
}
