% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFIT.R
\name{format_refdata}
\alias{format_refdata}
\title{classifier_Seurat}
\usage{
format_refdata(
  ref_seurat,
  ref_meta_batchID,
  ref_meta = NULL,
  ref_meta_labels = NULL,
  dataset.name = "atlas:"
)
}
\arguments{
\item{ref_seurat}{A Seurat object as reference}

\item{ref_meta_batchID}{a vector that defines splits/batches in the ref_seurat e.g. ref_seurat$myID}

\item{ref_meta}{the metadata of the ref object}

\item{dataset.name}{A name defaults 'atlas:'}

\item{ref_meta_labels(optional)}{charateristic vector of cell labels (default NULL)}
}
\description{
format_refdata takes seurat objects and returns appropriate list format
}
