% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFIT.R
\name{GenerateIntegratedReference}
\alias{GenerateIntegratedReference}
\title{GenerateIntegratedReference}
\usage{
GenerateIntegratedReference(
  seuratObj,
  batchIdCol = NULL,
  targetBatchSize = 5000,
  assay = "RNA",
  ngenes = 3000,
  saveFile = NULL,
  seed = 0,
  verbose = FALSE,
  minBatchSize = 100,
  selectVariableFeaturesFromWholeInput = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{batchIdCol}{If provided, data will be subset using this metadata field}

\item{targetBatchSize}{This is an alternative to batchIdCol. If batchIdCol is not provided, data will be randomly subset}

\item{assay}{The seurat assay holding count data}

\item{ngenes}{The number of variable genes to select}

\item{saveFile}{If provided, the integrated reference object will be saved to this file as RDS}

\item{seed}{The random seed}

\item{verbose}{Determined output verbosity}

\item{minBatchSize}{If any resulting batch has fewer than this many cells, an error is thrown}

\item{selectVariableFeaturesFromWholeInput}{If true, the variable genes will be calculated from the intact input object. If false, FindVariableFeatures() will be calculated independently on each subset seurat object, followed by Seurat::SelectIntegrationFeatures}
}
\description{
This function uses cFIT to create an integrated reference for the provided seurat object
}
