% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_FXs.R
\name{GetAtlasMetadata}
\alias{GetAtlasMetadata}
\title{Helper function for ConvertAtlasAndApplyMetadata.}
\usage{
GetAtlasMetadata(
  basepath = "~/Box/Bimber/BimberLab Shared/Experiments/RhesusAtlas/data/meta/",
  freezeV = "June232021_freeze/"
)
}
\value{
A data frame with the current atlas metadata.
}
\description{
This needs amending in the future, but for now it grabs the metadata on Box
}
\examples{
atlasObj <- GetAtlasMetadata()
}
