% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFIT.R
\name{GetCFitReference}
\alias{GetCFitReference}
\title{Get the integrated cFIT reference object for an atlas version}
\usage{
GetCFitReference(version = NULL)
}
\arguments{
\item{version}{The version string}
}
\description{
Return the integrated reference for the provided RIRA version
}
