% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{GetRiraCountMatrix}
\alias{GetRiraCountMatrix}
\title{Get the count data for a given RIRA version, as a Seurat object}
\usage{
GetRiraCountMatrix(version = NULL)
}
\arguments{
\item{version}{The version string. If not provided, defaults to the current version.}
}
\description{
Get the count data for a given RIRA version, as a Seurat object
}
