% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate.R
\name{GetScGateModel}
\alias{GetScGateModel}
\title{GetScGateModel}
\usage{
GetScGateModel(gateName)
}
\arguments{
\item{gateName}{The name of the gate to return. See GetAvailableScGates() for a list of known gates}
}
\description{
Returns the selected scGate model
}
