% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_FXs.R
\name{integrate_ref}
\alias{integrate_ref}
\title{integrate_ref  perform integration on reference}
\usage{
integrate_ref(genes, data.list)
}
\arguments{
\item{genes}{A set of genes usually top variable genes}

\item{data.list}{the output of format_refdata()}
}
\description{
perform integration on reference cFIT
}
