% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_FXs.R
\name{integrate_reflist_CCA}
\alias{integrate_reflist_CCA}
\title{integrate_reflist_CCA}
\usage{
integrate_reflist_CCA(RefLS = NULL, dims = 1:30, dims_umap = 1:20)
}
\arguments{
\item{RefLS}{The seurat object ref list see SplitObject()}

\item{dims}{CCA/PCA dims to integrate default 1:30}

\item{dims_umap}{CCA/PCA dims to run UMAP default 1:20}
}
\value{
list of 2 times Ref.anchors and Ref.integrated
}
\description{
integrate CCA given a list of ref Seurat object see SplitObject() of Seurat
}
