% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_Seurat.R
\name{InterpretModels}
\alias{InterpretModels}
\title{Interprets the feature importance of each model}
\usage{
InterpretModels(output_dir = "./classifiers", plot_type = "ratio")
}
\arguments{
\item{output_dir}{The output directory that TrainAllModels saved training data and models into}

\item{plot_type}{Argument to pass to model_parts(). Ratio or difference is recommended for large feature sets where the base model's AUC loss will be outside the plotting range.}
}
\description{
Interprets the feature importance of each model using DALEX and feature importance permuation
}
