% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModules.R
\name{PlotUcellCorrelation}
\alias{PlotUcellCorrelation}
\title{PlotUcellCorrelation}
\usage{
PlotUcellCorrelation(seuratObj, toCalculate, assayName = "RNA")
}
\arguments{
\item{seuratObj}{The seurat object}

\item{toCalculate}{A named list where each item is a character vector of genes}

\item{assayName}{The assay to use}
}
\value{
A list with moduleName and the spearman correlation matrix
}
\description{
This will plot the correlation between a UCell score and a gene set of interest
}
