% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{PredictTcellActivation}
\alias{PredictTcellActivation}
\title{Predicts T cell activation using sPLS derived components and a trained logistic model on transformed variates}
\usage{
PredictTcellActivation(seuratObj, model = NULL)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{model}{The trained sPLS model to use for prediction. This can be a file path to an RDS file, or a built-in model name.}
}
\value{
A Seurat object with the sPLS scores and predicted probabilities added to the metadata
}
\description{
Predicts T cell activation using a trained model
}
