% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{PredictTcellActivationUsingCustomModel}
\alias{PredictTcellActivationUsingCustomModel}
\title{Predicts T cell activation using sPLS derived components and a trained logistic model on transformed variates}
\usage{
PredictTcellActivationUsingCustomModel(seuratObj, modelName, modelFile)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{modelName}{A name to use for this model. This will be the prefix of output columns}

\item{modelFile}{The path to an RDS with this model.}
}
\value{
A Seurat object with the sPLSDA scores and predicted probabilities added to the metadata
}
\description{
Predicts T cell activation using a trained model
}
