% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{RecoverUnassignedCells}
\alias{RecoverUnassignedCells}
\title{RecoverUnassignedCells}
\usage{
RecoverUnassignedCells(
  seuratObj,
  classField = "RIRA_Immune_v2.cellclass",
  groupField = "ClusterNames_0.2",
  targetField = "RIRA_Immune_v2.cellclass.recovered",
  unassignedValue = "Unassigned",
  minClusterProp = 0.6
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{classField}{The name of the field holding the cell type call}

\item{groupField}{The field on which to group, such as the cluster ID}

\item{targetField}{The field that will store the result}

\item{unassignedValue}{The string value that denotes Unassigned cells}

\item{minClusterProp}{If at least this proportion of the group is one class, unassigned cells will be assigned as this class}
}
\value{
The updated seurat object
}
\description{
This step recovers unassigned cells by inspecting each cluster and assigning Unassigned cells to the match the cluster majority
}
