% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{RegisterGeneSet}
\alias{RegisterGeneSet}
\title{RegisterGeneSet}
\usage{
RegisterGeneSet(name, genes)
}
\arguments{
\item{name}{The name of the gene set}

\item{genes}{A character vector of gene names}
}
\description{
Registers a named gene set in the package gene-set registry.
}
