% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{RunCellTypist}
\alias{RunCellTypist}
\title{Run Celltypist}
\usage{
RunCellTypist(seuratObj)
}
\arguments{
\item{seuratObj}{The seurat object}
}
\description{
Runs celltypist on a seurat object and stores the calls as metadata
}
