% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate.R
\name{RunScGate}
\alias{RunScGate}
\title{Run scGate}
\usage{
RunScGate(
  seuratObj,
  model,
  min.cells = 30,
  assay = "RNA",
  pos.thr = 0.2,
  neg.thr = 0.2,
  ncores = 1,
  output.col.name = "is.pure",
  genes.blacklist = "default",
  doPlotUCellScores = TRUE,
  keep.ranks = FALSE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{model}{Either an scGate model, or a character that will be passed to GetScGateModel()}

\item{min.cells}{Passed directly to scGate::scGate. Stop iterating if fewer than this number of cells is left}

\item{assay}{Passed directly to scGate::scGate. Seurat assay to use}

\item{pos.thr}{Passed directly to scGate::scGate. Minimum UCell score value for positive signatures}

\item{neg.thr}{Passed directly to scGate::scGate. Maximum UCell score value for negative signatures}

\item{ncores}{Passed directly to scGate::scGate. Number of processors for parallel processing (requires future.apply)}

\item{output.col.name}{Passed directly to scGate::scGate. Column name with 'pure/impure' annotation}

\item{genes.blacklist}{Passed directly to scGate::scGate. Genes blacklisted from variable features. The default loads the list of genes in scGate::genes.blacklist.default; you may deactivate blacklisting by setting genes.blacklist=NULL}

\item{doPlotUCellScores}{If true, FeaturePlots will be created for each UCell score used in classification}

\item{keep.ranks}{Passed directly to keep.ranks}
}
\description{
Helper function to run scGate
}
