% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate.R
\name{RunScGateForModels}
\alias{RunScGateForModels}
\title{Run scGate for models}
\usage{
RunScGateForModels(
  seuratObj,
  modelNames,
  min.cells = 10,
  assay = "RNA",
  pos.thr = 0.13,
  neg.thr = 0.13,
  ncores = 1,
  genes.blacklist = "default",
  labelRename = NULL,
  dropAmbiguousConsensusValues = FALSE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{modelNames}{A vector of model names to run. They are assumed to be non-overlapping populations}

\item{min.cells}{Passed directly to scGate::scGate. Stop iterating if fewer than this number of cells is left}

\item{assay}{Passed directly to scGate::scGate. Seurat assay to use}

\item{pos.thr}{Passed directly to scGate::scGate. Minimum UCell score value for positive signatures}

\item{neg.thr}{Passed directly to scGate::scGate. Maximum UCell score value for negative signatures}

\item{ncores}{Passed directly to scGate::scGate. Number of processors for parallel processing (requires future.apply)}

\item{genes.blacklist}{Passed directly to scGate::scGate. Genes blacklisted from variable features. The default loads the list of genes in scGate::genes.blacklist.default; you may deactivate blacklisting by setting genes.blacklist=NULL}

\item{labelRename}{An optional list that maps the model name to the final label that should be used in the seurat object. for exmaple: list(Tcell = 'T_NK', NK = 'T_NK'), would re-label cells classified as either 'Tcell' or 'NK' by those models to one common label of T_NK}

\item{dropAmbiguousConsensusValues}{If true, any consensus calls that are ambiguous will be set to NA}
}
\description{
Helper function to run scGate, iterating the provided models and generating a consensus field
}
