% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{ScoreCellsWithSavedModel}
\alias{ScoreCellsWithSavedModel}
\title{Applies a trained model to get per-cell probabilities.}
\usage{
ScoreCellsWithSavedModel(
  seuratObj,
  model,
  fieldName,
  batchSize = 20000,
  assayName = "RNA"
)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{model}{Either the full filepath to a model RDS file, or the name of a built-in model.}

\item{fieldName}{The name of the field used to store the results in the seurat object}

\item{batchSize}{To conserve memory, data will be chunked into batches of at most this many cells}

\item{assayName}{The assay holding gene expression data}
}
\description{
Applies a trained model to get per-cell probabilities.
}
