% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentScores.R
\name{ScoreUsingSavedComponent}
\alias{ScoreUsingSavedComponent}
\title{ScoreUsingSavedComponent}
\usage{
ScoreUsingSavedComponent(
  seuratObj,
  componentOrName,
  fieldName,
  assayName = "RNA",
  layer = "data"
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{componentOrName}{Either a data frame with the columns 'feature' and 'weight', or the name of a previously saved component available in RIRA}

\item{fieldName}{The name of the field in which to save the resulting scores}

\item{assayName}{The assay to use}

\item{layer}{The layer to use}
}
\description{
Scores a seurat object using the gene weights from a saved component
}
