% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SetAtlasDir}
\alias{SetAtlasDir}
\title{Set the directory where atlas data reside}
\usage{
SetAtlasDir(folderPath)
}
\arguments{
\item{folderPath}{The atlas directory. Within the folder, there should be sub-folders by version.}
}
\description{
Sets the directory where atlas data reside.
}
