% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_Seurat.R
\name{TrainAllModels}
\alias{TrainAllModels}
\title{Wrapper function for TrainModel to train a suite of binary classifiers for each cell type present in the data}
\usage{
TrainAllModels(
  seuratObj,
  celltype_column,
  assay = "RNA",
  slot = "data",
  output_dir = "./classifiers",
  hyperparameter_tuning = F,
  learner = "classif.ranger",
  inner_resampling = "cv",
  outer_resampling = "cv",
  inner_folds = 4,
  inner_ratio = 0.8,
  outer_folds = 3,
  outer_ratio = 0.8,
  n_models = 20,
  n_cores = NULL,
  gene_list = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{celltype_column}{The metadata column containing the celltypes. One classifier will be created for each celltype present in this column.}

\item{assay}{SeuratObj assay containing the desired count matrix/metadata}

\item{slot}{Slot containing the count data. Should be restricted to counts, data, or scale.data.}

\item{output_dir}{The directory in which models, metrics, and training data will be saved.}

\item{hyperparameter_tuning}{Logical that determines whether or not hyperparameter tuning should be performed.}

\item{learner}{The mlr3 learner that should be used. Currently fixed to "classif.ranger" if hyperparameter tuning is FALSE. Otherwise, "classif.xgboost" and "classif.ranger" are supported.}

\item{inner_resampling}{The resampling strategy that is used for hyperparameter optimization. Holdout ("hout" or "holdout") and cross validation ("cv" or "cross-validation") are supported.}

\item{outer_resampling}{The resampling strategy that is used to determine overfitting. Holdout ("hout" or "holdout") and cross validation ("cv" or "cross-validation") are supported.}

\item{inner_folds}{The number of folds to be used for inner_resampling if cross-valdiation is performed.}

\item{inner_ratio}{The ratio of training to testing data to be used for inner_resampling if holdout resampling is performed.}

\item{outer_folds}{The number of folds to be used for outer_resampling if cross-valdiation is performed.}

\item{outer_ratio}{The ratio of training to testing data to be used for inner_resampling if holdout resampling is performed.}

\item{n_models}{The number of models to be trained during hyperparameter tuning. The model with the highest accuracy will be selected and returned.}

\item{n_cores}{If non-null, this number of workers will be used with future::plan}

\item{gene_list}{If non-null, the input count matrix will be subset to these features}

\item{verbose}{Whether or not to print the metrics data for each model after training.}
}
\description{
Wrapper function for TrainModel to train a suite of binary classifiers for each cell type present in the data
}
