% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFIT.R
\name{transfer_sigs}
\alias{transfer_sigs}
\title{transfer_sigs  perform CFITTransfer}
\usage{
transfer_sigs(genes, target_seurat, int.out)
}
\arguments{
\item{genes}{A set of genes usually top variable genes}

\item{target_seurat}{A query seurat object}

\item{int.out}{the output of integrate_ref()}
}
\description{
perform CFITTransfer
}
