% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_Seurat.R
\name{AssignCellType}
\alias{AssignCellType}
\title{Assigns celltype label based on probabilities in the metadata}
\usage{
AssignCellType(seuratObj, minimum_probability = 0.5, minimum_delta = 0.25)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{minimum_probability}{The minimum probability for a confident cell type assignment}

\item{minimum_delta}{The minimum difference in probabilities necessary to call one celltype over another.}
}
\description{
Assigns celltype label based on probabilities in the metadata
}
