% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{AssignCellType}
\alias{AssignCellType}
\title{Assigns celltype label based on probabilities in the metadata}
\usage{
AssignCellType(
  seuratObj,
  probabilityColumns,
  fieldName = "RIRA_Consensus",
  minimum_probability = 0.5,
  minimum_delta = 0.25
)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{probabilityColumns}{Any column ending with this value will be assumed to be a cell type probability column}

\item{fieldName}{The name of the metadata column to store the result}

\item{minimum_probability}{The minimum probability for a confident cell type assignment}

\item{minimum_delta}{The minimum difference in probabilities necessary to call one celltype over another.}
}
\description{
Assigns celltype label based on probabilities in the metadata
}
