% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModules.R
\name{CalculateUCellScores}
\alias{CalculateUCellScores}
\title{CalculateUCellScores}
\usage{
CalculateUCellScores(
  seuratObj,
  forceRecalculate = FALSE,
  seed = GetSeed(),
  ncores = 1,
  assayName = "RNA",
  storeRanks = TRUE,
  plotCor = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{forceRecalculate}{If true, UCell will always be re-run, even if the field is already present.}

\item{seed}{If non-null, set.seed() will be called prior to running UCell}

\item{ncores}{The number of core to use with UCell::AddModuleScore_UCell}

\item{assayName}{The assay from which to calculate UCell scores.}

\item{storeRanks}{Passed directly to UCell::AddModuleScore_UCell. Increases object size but makes future calculations quicker.}

\item{plotCor}{If true, a plot of correlations between the UCell score and each component gene will be shown}
}
\description{
This will calculate a handful of standardized UCell scores for a seurat object
}
