% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_FXs.R
\name{classifier_Seurat}
\alias{classifier_Seurat}
\title{classifier_Seurat}
\usage{
classifier_Seurat()
}
\arguments{
\item{seuratObj}{The seurat object where data will be added.}
}
\description{
calls seurat fx to classify a query dataset
}
