% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{CombineTcellActivationClasses}
\alias{CombineTcellActivationClasses}
\title{Combine T cell activation classes using custom logic}
\usage{
CombineTcellActivationClasses(
  seuratObj,
  modelName = "General",
  modelVersion = "v3",
  classMapping,
  outputFieldName = NULL,
  probabilityAggregation = "sum"
)
}
\arguments{
\item{seuratObj}{The Seurat Object containing T cell activation predictions}

\item{modelName}{The name of the model whose predictions should be combined. Default is "General".}

\item{modelVersion}{The version of the model. Default is "v3" for the General model.}

\item{classMapping}{A named list where names are the new combined class labels and values are 
character vectors of the original classes to combine. For example:
list("Activated" = c("Early_Activated", "Late_Activated"), "Resting" = c("Naive", "Memory"))}

\item{outputFieldName}{The name of the metadata column to store the combined classifications. 
If NULL, will use "{modelName}_Combined_Class_{modelVersion}".}

\item{probabilityAggregation}{How to aggregate probabilities for combined classes. Options are:
"sum" (default) - sum probabilities of constituent classes,
"max" - take maximum probability among constituent classes,
"mean" - take mean of probabilities of constituent classes.}
}
\value{
A Seurat object with the combined class assignments added to metadata
}
\description{
Takes a Seurat object with T cell activation predictions (from PredictTcellActivation) 
and combines classes based on user-defined logic. This is useful for collapsing fine-grained 
activation states into broader categories.
}
\examples{
\dontrun{
# Combine activation states into broader categories
seuratObj <- CombineTcellActivationClasses(
  seuratObj,
  classMapping = list(
    "Activated" = c("Early_Activated", "Late_Activated"),
    "Resting" = c("Naive", "Resting")
  )
)
}
}
