% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_Seurat.R
\name{.ParseMetricsFile}
\alias{.ParseMetricsFile}
\title{Helper function for printing information about the metrics files}
\usage{
.ParseMetricsFile(metrics_file)
}
\description{
Parses the .rds file saved during TrainAllModels and prints accuracy information
}
