% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate.R
\name{GetScGateModel}
\alias{GetScGateModel}
\title{GetScGateModel}
\usage{
GetScGateModel(modelName, allowSCGateDB = TRUE)
}
\arguments{
\item{modelName}{The name of the gate to return. See GetAvailableScGates() for a list of known gates}

\item{allowSCGateDB}{If true, this will search local models and the models provided by scGate::get_scGateDB()}
}
\description{
Returns the selected scGate model
}
