% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{MakePhenotypingDotPlot}
\alias{MakePhenotypingDotPlot}
\title{MakePhenotypingDotPlot}
\usage{
MakePhenotypingDotPlot(
  seuratObj,
  yField = "ClusterNames_0.2",
  scaled = T,
  gene_lists = c("Cytotoxicity", "EffectorCytokines")
)
}
\arguments{
\item{yField}{The grouping variable used to calculate the average expression of genes and the y axis of the DotPlot.}

\item{scaled}{A boolean defining whether to color dots by scaled expression or unscaled expression.}

\item{gene_lists}{A vector of gene lists (defined by .RegisterGeneSet) to be queried and their genes be plotted.}
}
\description{
Creates a DotPlot using custom gene sets and attempts to coarsely group gene sets by cell type.
}
