% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFIT.R
\name{Perform_CFITTransfer}
\alias{Perform_CFITTransfer}
\title{Perform CFITTransfer}
\usage{
Perform_CFITTransfer(
  seuratObj,
  integratedReference = GetCFitReference(),
  verbose = TRUE,
  seed = 0,
  assay = "RNA"
)
}
\arguments{
\item{seuratObj}{The input seurat object}

\item{integratedReference}{The integrated reference from cFIT}

\item{verbose}{Controls the verbosity of the output}

\item{seed}{The random reed}

\item{assay}{The name of the assay holding count data}
}
\description{
Perform CFITTransfer
}
