% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_Seurat.R
\name{PredictCellTypeProbability}
\alias{PredictCellTypeProbability}
\title{Applies trained models to get celltype probabilities}
\usage{
PredictCellTypeProbability(
  seuratObj,
  models_dir = "./classifiers/models",
  iterative = FALSE,
  iterations = 100,
  assayName = "RNA"
)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{models_dir}{the directory containing the models produced by TrainAllModels}

\item{iterative}{Logical determining whether or not the data should be chunked to save memory or not.}

\item{iterations}{Integer determing how many chunks the data should be split into.}

\item{assayName}{The assay holding gene expression data}
}
\description{
Applies trained models to get celltype probabilities.
}
