% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_Seurat.R
\name{PredictCellTypeProbability}
\alias{PredictCellTypeProbability}
\title{Applies trained models to get celltype probabilities}
\usage{
PredictCellTypeProbability(
  seuratObj,
  models_dir = "./classifiers/models",
  batchSize = 20000,
  assayName = "RNA"
)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{models_dir}{the directory containing the models produced by TrainAllModels}

\item{batchSize}{To conserve memory, data will be chunked into batches of at most this many cells}

\item{assayName}{The assay holding gene expression data}
}
\description{
Applies trained models to get celltype probabilities.
}
