% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{RunCellTypist}
\alias{RunCellTypist}
\title{Run Celltypist}
\usage{
RunCellTypist(
  seuratObj,
  modelName = "Immune_All_Low.pkl",
  extraArgs = c("--majority-voting", "--mode", "prob_match", "--p-thres", 0.5),
  assayName = "RNA",
  columnPrefix = NULL,
  convertAmbiguousToNA = FALSE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{modelName}{The model name or path to celltypist model}

\item{extraArgs}{An optional list of additional arguments passed directly on the command line to cell typist}

\item{assayName}{The name of the assay to use. Others will be dropped}

\item{columnPrefix}{A prefix that will be added to the beginning of the resulting columns, added the seurat@meta.data}

\item{convertAmbiguousToNA}{If true, any values for majority_voting with commas (indicating they are ambiguous) will be converted to NA}
}
\description{
Runs celltypist on a seurat object and stores the calls as metadata
}
