% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{RunCellTypist}
\alias{RunCellTypist}
\title{Run Celltypist}
\usage{
RunCellTypist(
  seuratObj,
  modelName = "Immune_All_Low.pkl",
  pThreshold = 0.5,
  minProp = 0,
  useMajorityVoting = TRUE,
  mode = "prob_match",
  extraArgs = c("--mode", mode, "--p-thres", pThreshold, "--min-prop", minProp),
  assayName = Seurat::DefaultAssay(seuratObj),
  columnPrefix = NULL,
  maxAllowableClasses = 6,
  minFractionToInclude = 0.01,
  minCellsToRun = 200,
  maxBatchSize = 6e+05,
  retainProbabilityMatrix = FALSE,
  runCelltypistUpdate = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{modelName}{The model name or path to celltypist model}

\item{pThreshold}{By default, this would be passed to the --p-thres argument. However, if you also provide extraArgs, this is ignored.}

\item{minProp}{By default, this would be passed to the --min-prop argument. However, if you also provide extraArgs, this is ignored.}

\item{useMajorityVoting}{If true, the celltypist --majority-voting option will be added.}

\item{mode}{Passed directly to the celltypist --mode argument. Either best_match or prob_match}

\item{extraArgs}{An optional list of additional arguments passed directly on the command line to cell typist}

\item{assayName}{The name of the assay to use. Others will be dropped}

\item{columnPrefix}{A prefix that will be added to the beginning of the resulting columns, added the seurat@meta.data}

\item{maxAllowableClasses}{Celltypist can assign a cell to many classes, creating extremely long labels. Any cell with more than this number of labels will be set to NA}

\item{minFractionToInclude}{If non-null, any labels with fewer than this fraction of cells will be set to NA.}

\item{minCellsToRun}{If the input seurat object has fewer than this many cells, NAs will be added for all expected columns and celltypist will not be run.}

\item{maxBatchSize}{If more than this many cells are in the object, it will be split into batches of this size and run in serial.}

\item{retainProbabilityMatrix}{If true, the celltypist probability_matrix with per-class probabilities will be stored in meta.data}

\item{runCelltypistUpdate}{If true, --update-models will be run for celltypist prior to scoring cells.}
}
\description{
Runs celltypist on a seurat object and stores the calls as metadata
}
