% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{RunCellTypist}
\alias{RunCellTypist}
\title{Run Celltypist}
\usage{
RunCellTypist(
  seuratObj,
  modelName = "Immune_All_Low.pkl",
  pThreshold = 0.5,
  minProp = 0,
  extraArgs = c("--majority-voting", "--mode", "prob_match", "--p-thres", pThreshold,
    "--min-prop", minProp),
  assayName = "RNA",
  columnPrefix = NULL,
  convertAmbiguousToNA = FALSE,
  maxAllowableClasses = 6,
  minFractionToPlot = 0.01
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{modelName}{The model name or path to celltypist model}

\item{pThreshold}{By default, this would be passed to the --p-thres argument. However, if you also provide extraArgs, this is ignored.}

\item{minProp}{By default, this would be passed to the --min-prop argument. However, if you also provide extraArgs, this is ignored.}

\item{extraArgs}{An optional list of additional arguments passed directly on the command line to cell typist}

\item{assayName}{The name of the assay to use. Others will be dropped}

\item{columnPrefix}{A prefix that will be added to the beginning of the resulting columns, added the seurat@meta.data}

\item{convertAmbiguousToNA}{If true, any values for majority_voting with commas (indicating they are ambiguous) will be converted to NA}

\item{maxAllowableClasses}{Celltypist can assign a cell to many classes, creating extremely long labels. Any cell with more than this number of labels will be set to NA}

\item{minFractionToPlot}{If non-null, any labels with fewer than this fraction of cells will be set to 'Low Freq' for the purposes of generating the summary barplot.}
}
\description{
Runs celltypist on a seurat object and stores the calls as metadata
}
