% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{TrainCellTypist}
\alias{TrainCellTypist}
\title{Train Celltypist}
\usage{
TrainCellTypist(seuratObj, labelField, modelFile)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{labelField}{The field in seuratObj@meta.data holding the labels for training}

\item{modelFile}{The path to save the model}
}
\description{
Train celltypist using the data in a seurat object and saves the resulting model
}
