% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypist.R
\name{TrainCellTypist}
\alias{TrainCellTypist}
\title{Train Celltypist}
\usage{
TrainCellTypist(seuratObj, labelField, modelFile, minCellsPerClass = 20)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{labelField}{The field in seuratObj@meta.data holding the labels for training}

\item{modelFile}{The path to save the model}

\item{minCellsPerClass}{If provided, any classes (and corresponding cells) with fewer than this many cells will be dropped from the training data}
}
\description{
Train celltypist using the data in a seurat object and saves the resulting model
}
